requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
pydantic-settings
python-jose[cryptography]
python-multipart
python-dotenv
requests
aiosqlite
passlib[bcrypt]
